requirements.txt
# Telegram Bot API
python-telegram-bot==20.7

# 异步数据库
aiosqlite==0.19.0

# 异步HTTP客户端
aiohttp==3.9.1

# 环境变量管理
python-dotenv==1.0.0

# 文件处理
Pillow==10.1.0

# 日期时间处理
python-dateutil==2.8.2

# 日志格式化
colorlog==6.8.0

# 异步文件操作
aiofiles==23.2.1

# 系统信息
psutil==5.9.6

# 网络请求
requests==2.31.0

# 配置文件处理
PyYAML==6.0.1

# 数据验证
pydantic==2.5.0

# 异步任务队列
celery==5.3.4

# 缓存支持
redis==5.0.1

# 数据库迁移
alembic==1.13.1

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1

# 代码质量
black==23.11.0
flake8==6.1.0
mypy==1.7.1